// src/types/Employee.ts
export interface Employee {
  emp_id: string;
  resource_name: string;
  prj_align: string;
  core_alignment: string;
  core_team: string;
  job_title: string;
  role_type: string;
  status: string;
  base_location: string;
  email_id: string;
  hire_date: string;
  term_date?: string;
  modified_by: string;
  modified_at: string;
  vendor: string;
  contact_number: string;
  team_name: string;
  manager_name: string;
  secondary_team: string;
}